fix: clean up code and minor fixes
